Fixed MEF support / added clean function
Added support for a custom clean function per file format. This method
will automatically be called when the object is destroyed or the
cleanup function is called.

MEF file format now chaches the MEF header as a temporary file in the
tempdir. these files are removed when the object is destroyed